pyproject.toml
[project]
name = "temp-btp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "catboost>=1.2.10",
    "ctgan>=0.12.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "kagglehub>=1.0.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
    "xgboost>=3.2.0",
]
